Split bench result types from parser
