+++ src/core/extension/bench/parsing.rs
@@ -56,225 +56,21 @@
 use std::collections::BTreeMap;
 use std::path::Path;
 
-use serde::{Deserialize, Serialize};
-
 use crate::core::error::{Error, Result};
-use crate::core::finding::HomeboyFinding;
-use crate::core::observation::timeline::{
-    reporting_timeline, summarize_spans, ObservationEvent, ObservationSpanDefinition,
-    ObservationSpanResult,
-};
+use crate::core::observation::timeline::{reporting_timeline, summarize_spans};
 
-use super::artifact::BenchArtifact;
 use super::artifact_validation;
-use super::diagnostic::BenchDiagnostic;
-use super::distribution::BenchRunDistribution;
-use super::gate::{BenchGate, BenchGateResult};
-use super::metric_policy_preset::{expand_metric_policy_presets, BenchMetricPolicyPreset};
-use super::phase_events::{
-    evaluate_phase_events, BenchPhaseEvent, BenchPhaseFailureClassification, BenchPhaseSummary,
+use super::metric_policy_preset::expand_metric_policy_presets;
+use super::phase_events::evaluate_phase_events;
+
+pub use super::result_types::{
+    BenchMemory, BenchMetricDirection, BenchMetricPhase, BenchMetricPolicy, BenchMetrics,
+    BenchResults, BenchRunExecution, BenchRunMetadata, BenchRunSnapshot, BenchRunnerMetadata,
+    BenchScenario, BenchWorkloadMetadata, RegressionTest,
 };
-
-fn default_true() -> bool {
-    true
-}
-
-fn is_true(value: &bool) -> bool {
-    *value
-}
-
-/// Full bench run output from an extension script.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub struct BenchResults {
-    pub component_id: String,
-    pub iterations: u64,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub run_metadata: Option<BenchRunMetadata>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metadata: BTreeMap<String, serde_json::Value>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metric_groups: BTreeMap<String, BTreeMap<String, f64>>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub timeline: Vec<serde_json::Value>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub span_definitions: BTreeMap<String, serde_json::Value>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub diagnostics: Vec<BenchDiagnostic>,
-    /// Structured lifecycle events emitted by the bench runner.
-    ///
-    /// Runners can append these as long phases progress so persisted bench
-    /// evidence can distinguish dependency setup, runtime startup, workload
-    /// execution, artifact collection, or any other runner-defined phase.
-    /// Extension-specific details belong in `payload`.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub phase_events: Vec<BenchPhaseEvent>,
-    /// Derived phase rollup computed from `phase_events`.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub phase_summaries: Vec<BenchPhaseSummary>,
-    /// Derived classification for the first timeout/failure phase.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub failure_classification: Option<BenchPhaseFailureClassification>,
-    #[serde(
-        default,
-        deserialize_with = "super::budget_findings::deserialize_budget_findings",
-        skip_serializing_if = "Vec::is_empty"
-    )]
-    pub budget_findings: Vec<HomeboyFinding>,
-    pub scenarios: Vec<BenchScenario>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metric_policies: BTreeMap<String, BenchMetricPolicy>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metric_policy_presets: BTreeMap<String, BenchMetricPolicyPreset>,
-}
-
-/// Homeboy-owned reproducibility metadata for a bench invocation.
-///
-/// Extension runners are not required to emit this block. Homeboy stamps it
-/// after parsing so stored bench artifacts explain what ran without requiring
-/// each language runner to duplicate CLI/runtime bookkeeping.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
-#[serde(deny_unknown_fields)]
-pub struct BenchRunMetadata {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub homeboy_version: Option<String>,
-    pub started_at: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub shared_state: Option<String>,
-    pub iterations: u64,
-    #[serde(flatten)]
-    pub execution: BenchRunExecution,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub warmup_iterations: Option<u64>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub selected_scenarios: Vec<String>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub env_overrides: BTreeMap<String, String>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub workloads: Vec<BenchWorkloadMetadata>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub runner: Option<BenchRunnerMetadata>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub diagnostics: Vec<BenchDiagnostic>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Default)]
-#[serde(deny_unknown_fields)]
-pub struct BenchRunExecution {
-    pub runs: u64,
-    pub concurrency: u32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub struct BenchWorkloadMetadata {
-    pub id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sha256: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub struct BenchRunnerMetadata {
-    pub extension: String,
-    pub path: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source_revision: Option<String>,
-}
-
-/// One scenario's measurements.
-///
-/// Scenario-level unknown keys are accepted to keep the contract
-/// forward-compatible: a runner can emit extra metadata (tags, warmup
-/// counts, environment info) without breaking parsers.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct BenchScenario {
-    pub id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub file: Option<String>,
-    /// Scenario origin. Dispatchers use `in_tree` for component-owned
-    /// workloads and `rig` for out-of-tree workloads supplied by a rig spec.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-    /// Declared default iteration count. List-only discovery uses this to
-    /// expose runner defaults without executing the workload.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub default_iterations: Option<u64>,
-    /// Freeform scenario labels supplied by extension runners.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub tags: Vec<String>,
-    pub iterations: u64,
-    pub metrics: BenchMetrics,
-    /// Optional grouped numeric metrics for secondary metric families.
-    ///
-    /// Flat `metrics` remains the primary backwards-compatible contract.
-    /// Runners can opt into grouped metrics when a scenario naturally emits
-    /// related values (for example phase timings or tool-call stats) without
-    /// flattening those groups in the source JSON.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metric_groups: BTreeMap<String, BTreeMap<String, f64>>,
-    /// Optional scenario timeline for causal evidence.
-    ///
-    /// Bench scenarios use the shared observation timeline contract so
-    /// runners can emit phase/span evidence without flattening everything
-    /// into metrics. Homeboy derives `span_results` from this timeline when
-    /// `span_definitions` are present.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub timeline: Vec<ObservationEvent>,
-    /// Optional span definitions over `source.event` timeline keys.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub span_definitions: Vec<ObservationSpanDefinition>,
-    /// Computed span outcomes, populated by Homeboy after parsing.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub span_results: Vec<ObservationSpanResult>,
-    /// Scenario-level semantic gates. Unlike metric policies, gates are
-    /// correctness checks: any failure invalidates the scenario even if
-    /// timing metrics improved.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub gates: Vec<BenchGate>,
-    /// Computed gate outcomes, populated by Homeboy after metrics are
-    /// parsed and aggregated.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub gate_results: Vec<BenchGateResult>,
-    /// Scenario-level categorical metadata emitted by benchmark workloads.
-    ///
-    /// Numeric values belong in `metrics`; this freeform object is for
-    /// labels and feature flags that are useful when aggregating persisted
-    /// observation runs, such as design fingerprints or workload choices.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metadata: BTreeMap<String, serde_json::Value>,
-    /// Scenario pass/fail status after semantic gates are evaluated.
-    #[serde(default = "default_true", skip_serializing_if = "is_true")]
-    pub passed: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub memory: Option<BenchMemory>,
-    /// Optional artifact pointers produced by the scenario.
-    ///
-    /// Homeboy preserves paths/URLs and metadata but does not upload, retain,
-    /// or diff artifact contents. Consumers can correlate artifacts by
-    /// scenario, rig, and run without scraping logs or side-channel files.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub artifacts: BTreeMap<String, BenchArtifact>,
-    /// Workload-emitted diagnostics for this scenario.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub diagnostics: Vec<BenchDiagnostic>,
-    /// Per-run raw metric snapshots when `homeboy bench --runs N` is used.
-    /// Omitted for the default `--runs 1` path so existing envelopes keep
-    /// their exact shape.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub runs: Option<Vec<BenchRunSnapshot>>,
-    /// Cross-run distribution stats keyed by metric name. Omitted for the
-    /// default `--runs 1` path.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub runs_summary: Option<BTreeMap<String, BenchRunDistribution>>,
-}
 
 /// Derive scenario span results from the shared observation timeline contract.
-pub fn evaluate_spans(results: &mut BenchResults) {
+fn evaluate_spans(results: &mut BenchResults) {
     for scenario in &mut results.scenarios {
         if scenario.span_definitions.is_empty() {
             continue;
@@ -282,141 +78,6 @@ pub fn evaluate_spans(results: &mut BenchResults) {
         let timeline = reporting_timeline(&scenario.timeline);
         scenario.span_results = summarize_spans(&timeline, &scenario.span_definitions);
     }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct BenchRunSnapshot {
-    pub metrics: BenchMetrics,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metric_groups: BTreeMap<String, BTreeMap<String, f64>>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub timeline: Vec<ObservationEvent>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub span_definitions: Vec<ObservationSpanDefinition>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub span_results: Vec<ObservationSpanResult>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub memory: Option<BenchMemory>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub artifacts: BTreeMap<String, BenchArtifact>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub diagnostics: Vec<BenchDiagnostic>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
-pub struct BenchMetrics {
-    #[serde(flatten)]
-    pub values: BTreeMap<String, f64>,
-    /// Raw per-iteration samples for variance-aware metrics.
-    ///
-    /// `values` remains the single-point summary contract; distributions
-    /// are opt-in data used by variance-aware regression checks.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub distributions: BTreeMap<String, Vec<f64>>,
-}
-
-impl BenchMetrics {
-    pub fn get(&self, key: &str) -> Option<f64> {
-        self.values.get(key).copied()
-    }
-
-    pub fn distribution(&self, key: &str) -> Option<&[f64]> {
-        self.distributions.get(key).map(Vec::as_slice)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub struct BenchMetricPolicy {
-    pub direction: BenchMetricDirection,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub regression_threshold_percent: Option<f64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub regression_threshold_absolute: Option<f64>,
-    /// True when this metric is expected to vary between iterations.
-    ///
-    /// Variance-aware metrics must emit a matching `metrics.distributions`
-    /// entry so regression checks can compare distribution shape instead
-    /// of a single summary point.
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub variance_aware: bool,
-    /// Minimum sample count needed for a meaningful variance-aware run.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub min_iterations_for_variance: Option<u64>,
-    /// Statistical test used for variance-aware regression detection.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub regression_test: Option<RegressionTest>,
-    /// Optional measurement-phase tag.
-    ///
-    /// Phase is **metadata only**: it does not affect regression math
-    /// (cold and warm metrics use the same `direction` /
-    /// `regression_threshold_*` fields), but it lets report renderers
-    /// group metrics by phase so cold-start numbers don't mix with
-    /// steady-state numbers in the same row of a diff table.
-    ///
-    /// Backwards-compatible: pre-existing JSON without `phase`
-    /// deserializes as `None` and round-trips unchanged.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub phase: Option<BenchMetricPhase>,
-}
-
-fn is_false(value: &bool) -> bool {
-    !*value
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-pub enum BenchMetricDirection {
-    #[serde(rename = "lower_is_better", alias = "lower")]
-    LowerIsBetter,
-    #[serde(rename = "higher_is_better", alias = "higher")]
-    HigherIsBetter,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum RegressionTest {
-    /// Legacy single-point threshold comparison.
-    PointDelta,
-    /// Non-parametric rank test, useful when distributions are not Normal.
-    MannWhitneyU,
-    /// Distribution-shape test sensitive to CDF shifts.
-    KolmogorovSmirnov,
-}
-
-/// Measurement-phase tag for a metric.
-///
-/// A bench run can mix one-time setup costs (process spawn, WASM boot,
-/// dependency install) with steady-state per-iteration costs. Without
-/// this tag every metric ends up in one flat alphabetical list and a
-/// 3500ms cold-boot sits next to a 12ms warm request as though they
-/// were comparable. Phase tagging lets the report renderer group cold
-/// metrics first, warm metrics second, amortized last, so the diff
-/// reads as the actually-useful story instead of a flat dump.
-///
-/// Phase is **opt-in**: pre-existing policies without a `phase` field
-/// stay untagged and render identically to today.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-#[serde(rename_all = "lowercase")]
-pub enum BenchMetricPhase {
-    /// One-time setup cost (process spawn, WASM boot, dependency
-    /// install). First iteration only; subsequent iterations don't pay
-    /// this cost unless the dispatcher restarts the substrate between
-    /// iterations.
-    Cold,
-    /// Steady-state per-iteration cost after warmup. The metric the
-    /// user sees on every request after the first.
-    Warm,
-    /// Synthetic blend, e.g. `(cold + N * warm) / N` for some N.
-    /// Useful for "what does the user see on first page-load"
-    /// framing where one cold request is amortized over a small
-    /// burst of warm follow-ups.
-    Amortized,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(deny_unknown_fields)]
-pub struct BenchMemory {
-    pub peak_bytes: u64,
 }
 
 /// Read and parse a `$HOMEBOY_BENCH_RESULTS_FILE` written by an extension.
